New abstract class GitHostingProviderBehavior
Parent Class for Git Hosting Provider Behavior, factors out duplicated
code from GitHubBehavior and GitLabBehavior classes (which have been
reduced to a single property override).
